Weird bug where stepper is replaced and then the first one is
restored, but it continues acting like the second
stepper, even though when printed the object shows as the first,
and loaded, stepper.

To replicate, click the icon in the toolbar to start stuff up
and show the first fragment, then highlight text and select, in
the context menu, "Read selected text", then press the icon again.
The "selected text" will stay displayed and playing will play
the "selected text" despite the fact that the console logs will
show the "full page" text stepper object.